support training for Apple M-series chip

+++ models/A2C_QPL.py
@@ -274,7 +274,7 @@ class A2C_QPL(object):
             R = r + 0.95 * R # R: culumative rewards for t to T
             returns.insert(0, R) # Evaluate the R and keep original order
 
-        returns = torch.tensor(returns).to(self.device)
+        returns = torch.tensor(returns, dtype=torch.float32).to(self.device)
         # Normalized returns
         returns = (returns - returns.mean()) / (returns.std() + eps)
